backend/requirements.txt
alembic==1.14.0
bcrypt==4.2.1
blinker==1.9.0
boto3==1.35.76
botocore==1.35.76
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
decouple==0.0.7
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.8
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
